Improved the AP framework (with magic)
The AP backend is now recursive and based around functions rather than
classes, which makes it pretty much infinitely more flexible. It
accommodates structures of arbitrary form and depth and could easily be
extended to QVectors and other numerically indexed containers.

The one potential drawback is that param setting is rather slow, as it
recurses through nested lambda functions. This could be remedied in
principle by setting via direct reference (obtained recursively, used in
a single call), but at the cost of stability.